GrammarMacroGenerator added.
+ Simple L-system (context free grammar) implementation done.
+ MacroGenerator structure slightly improved.
+ Song duration property added (to be used later).

+++ Assets/barelyMusician/Core/Scripts/Composition/Base/Generation/Algorithms/Markov Chain/MarkovChain.cs
@@ -45,7 +45,7 @@ namespace BarelyAPI
 
         public void GenerateNextState()
         {
-            double p = RandomNumber.NextFloat();
+            float p = RandomNumber.NextFloat();
 
             float cumulative = 0.0f;
             for (int i = 0; i < states.GetLength(1); ++i)
